Added functionality for add/remove items
